+++ DuckovCustomModel.Core/Managers/IAnimatorParameterUpdater.cs
@@ -1,7 +1,0 @@
-namespace DuckovCustomModel.Core.Managers
-{
-    public interface IAnimatorParameterUpdater
-    {
-        void UpdateParameters(object control, object context);
-    }
-}